pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = "Task Project"
authors = ["Burhan0014 <testleadz0014@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-crontab = "^0.7.1"
private-package = {version = "0.1.8", allow-prereleases = true, source = "packagr"}
pytest-factoryboy = "^2.7.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
pytest-cov = "^5.0.0"
coverage = "^7.6.0"
pytest-mock = "^3.14.0"

[tool.poetry.group.main.dependencies]
Django = "3.2.5"
django-debug-toolbar= "4.3.0"
djangorestframework= "3.15.1"
social-auth-app-django= "5.4.1"
psycopg2-binary= "2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "7.4"
pytest-cov = "^5.0.0"
coverage = "^7.6.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "gcs.settings"
addopts = """
    --reuse-db
    --ds=gcs.settings
"""